Implement js json schema validation :sparkles: #90
