+++ src/ISADotNet.Json/ISADotNet.Json.fsproj
@@ -4,6 +4,12 @@
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
   </PropertyGroup>
   <ItemGroup>
+    <EmbeddedResource Include="Validation\JsonValidation.js">
+		<CopyToOutputDirectory>Always</CopyToOutputDirectory>
+	</EmbeddedResource>
+    <Compile Include="Validation/ValidationResult.fs" />
+    <Compile Include="Validation\Fable.fs" />
+    <Compile Include="Validation/JsonSchemaValidation.fs" />
     <Compile Include="ConverterOptions.fs" />
     <Compile Include="Decode.fs" />
     <Compile Include="GEncode.fs" />
@@ -21,6 +27,7 @@
     <Compile Include="Investigation.fs" />
   </ItemGroup>
   <ItemGroup>
+	<PackageReference Include="NJsonSchema" Version="10.8.0" />
     <PackageReference Include="Thoth.Json" Version="10.1.0" />
     <PackageReference Include="Thoth.Json.Net" Version="11.0.0" />
   </ItemGroup>
