+++ tests/ISADotNet.Json.Tests/Main.fs
@@ -9,7 +9,7 @@ open Expecto
 #endif
 let all = testSequenced <| testList "All" [
     ISAJsonTests.main
-    //JSchemaValidationTests.main
+    JSchemaValidationTests.main
 ]
